+++ packages/stream_chat_flutter/lib/src/message_input/quoted_message_widget.dart
@@ -354,7 +354,6 @@ class _UrlAttachment extends StatelessWidget {
 
 class _VideoAttachmentThumbnail extends StatefulWidget {
   const _VideoAttachmentThumbnail({
-    super.key,
     required this.attachment,
   });