+++ packages/stream_chat_flutter/lib/src/video/video_thumbnail_image.dart
@@ -97,8 +97,9 @@ class _StreamVideoThumbnailImageState extends State<StreamVideoThumbnailImage> {
                     );
               }
               if (!snapshot.hasData) {
-                return Container(
-                  constraints: const BoxConstraints.expand(),
+                return SizedBox(
+                  height: double.maxFinite,
+                  width: double.maxFinite,
                   child: widget.placeholderBuilder?.call(context) ??
                       Shimmer.fromColors(
                         baseColor: _streamChatTheme.colorTheme.disabled,
@@ -106,16 +107,22 @@ class _StreamVideoThumbnailImageState extends State<StreamVideoThumbnailImage> {
                         child: Image.asset(
                           'images/placeholder.png',
                           fit: BoxFit.cover,
+                          height: widget.constraints?.maxHeight,
+                          width: widget.constraints?.maxWidth,
                           package: 'stream_chat_flutter',
                         ),
                       ),
                 );
               }
-              return Image.memory(
-                snapshot.data!,
-                fit: widget.fit,
-                height: widget.constraints?.maxHeight ?? double.infinity,
-                width: widget.constraints?.maxWidth ?? double.infinity,
+              return SizedBox(
+                height: double.maxFinite,
+                width: double.maxFinite,
+                child: Image.memory(
+                  snapshot.data!,
+                  fit: widget.fit,
+                  height: widget.constraints?.maxHeight ?? double.infinity,
+                  width: widget.constraints?.maxWidth ?? double.infinity,
+                ),
               );
             },
           ),