+++ packages/stream_chat_persistence/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Upcoming
+
+- Added support for `Channel.ownCapabilities`
+
 ## 4.1.0
 
 🔄 Changed